Fake Central EGA has now containers on its own
