+++ docker/images/cega_mq/Dockerfile
@@ -1,0 +1,5 @@
+FROM rabbitmq:management
+LABEL maintainer "Frédéric Haziza, NBIS"
+
+COPY rabbitmq.config /etc/rabbitmq/rabbitmq.config
+COPY rabbitmq.json /etc/rabbitmq/defs.json
