+++ docker/images/cega_mq/rabbitmq.config
@@ -1,0 +1,11 @@
+%% -*- mode: erlang -*-
+%%
+[{rabbit,[{loopback_users, [ ] },
+	  {default_vhost, "/"},
+	  {default_user,  "guest"},
+	  {default_pass,  "guest"},
+	  {default_permissions, [".*", ".*",".*"]},
+	  {default_user_tags, [administrator]},
+	  {disk_free_limit, "1GB"}]},
+ {rabbitmq_management, [ {load_definitions, "/etc/rabbitmq/defs.json"} ]}
+].
