+++ docker/images/cega_mq/rabbitmq.json
@@ -1,0 +1,25 @@
+{"rabbit_version":"3.6.11",
+ "users":[{"name":"guest", "password_hash":"4tHURqDiZzypw0NTvoHhpn8/MMgONWonWxgRZ4NXgR8nZRBz", "hashing_algorithm":"rabbit_password_hashing_sha256", "tags":"administrator"},
+	  {"name":"test" , "password_hash":"xa/aoJdq+EeI+/o5/IacnDLm/KyBim6uGBlU2+6nZrb6bIer", "hashing_algorithm":"rabbit_password_hashing_sha256", "tags":"administrator"}],
+ "vhosts":[{"name":"/"}, {"name":"test"}],
+ "permissions":[{"user":"guest", "vhost":"/", "configure":".*", "write":".*", "read":".*"},
+		{"user":"test" , "vhost":"/", "configure":".*", "write":".*", "read":".*"},
+		{"user":"guest", "vhost":"test", "configure":".*", "write":".*", "read":".*"},
+		{"user":"test" , "vhost":"test", "configure":".*", "write":".*", "read":".*"}],
+ "parameters":[],
+ "global_parameters":[{"name":"cluster_name", "value":"rabbit@localhost"}],
+ "policies":[],
+ "queues":[{"name":"sweden.v1.commands.file"     , "vhost":"test", "durable":true, "auto_delete":false, "arguments":{}},
+	   {"name":"sweden.v1.commands.account"  , "vhost":"test", "durable":true, "auto_delete":false, "arguments":{}},
+	   {"name":"sweden.v1.commands.completed", "vhost":"test", "durable":true, "auto_delete":false, "arguments":{}},
+	   {"name":"sweden.v1.commands.user"     , "vhost":"test", "durable":true, "auto_delete":false, "arguments":{}}],
+ "exchanges":[{"name":"localega.v1", "vhost":"test", "type":"topic", "durable":true, "auto_delete":false, "internal":false, "arguments":{}}],
+ "bindings":[{"source":"localega.v1", "vhost":"test", "destination_type":"queue", "arguments":{},
+	      "destination":"sweden.v1.commands.file", "routing_key":"sweden.file"},
+	     {"source":"localega.v1", "vhost":"test", "destination_type":"queue", "arguments":{},
+	      "destination":"sweden.v1.commands.completed", "routing_key":"sweden.file.completed"},
+	     {"source":"localega.v1", "vhost":"test", "destination_type":"queue", "arguments":{},
+	      "destination":"sweden.v1.commands.user", "routing_key":"sweden.user"},
+	     {"source":"localega.v1", "vhost":"test", "destination_type":"queue", "arguments":{},
+	      "destination":"sweden.v1.commands.account", "routing_key":"sweden.user.account"}]
+}
