--- src/auth/fake_cega/users/fred.yml
@@ -1,4 +1,0 @@
----
-password_hash: $1$xyz$sx8gPI05DJdJe4MJx5oXo0
-pubkey: ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQCyKoPimxBFwUYx1SDYZMV8x/IVdisShv0kvcJ3SRXny5vR7wTGU1gz6jaoM1mG+i+5NszJKxsEj4/HXHYVcOWithVw81cLuaA4nlSlQEwLqgPSBkUvRV29BHaeIBsbNY9pJOOr/PczTo5gKMFER69VbNmwixVHyAtngZdMjdcp0PezbKQ3xmSmYuKH5+CKzx69r/saEyoLLr5enQ0KkJJyNyVQRZb0Qoxjz0Qtno0c0NJzm8ivLp9G6y0p+/aQ/K/BUhsIgwgY5aSEePTU2iQ4tibgVu/1+XR7Y+7F9PIEXcQ6tLRQTQPLNx1ikZy3lWqKNOU2Dp9nR0QU69LUjEVN8jKjiOIu83KMIUZgFFmTZB4VoppjLZey1pIQuUeiXrxQcMKnOMRabHgh3M0bzlEwUVOD2jspSlLooE4J4gh4EMmCCvDwawe2pkBJMfZYk2nyWQDDiFpazQvEKBg5QW/WLvMLOPfpeEdNJLj6HRARAJaDhIcXFxIaDLmVDaoOUlUN9padxFNQtIWnw/yE8livGSNPM3DSGJ+/fZCQQouvWlppg4kV8HDt/NwMwPUqnWXy++tNbpo2QxPeyCcA6lruRaq944aO+9rafnuWYC6coUmJNoCoNmuB3W1aeeAsuoJx0zt0LhVG/L3/Ea3fQDmECMXPArutX2j37q4E8xMiFw== daz.admin@lega.sftp
-expiration: 1 month
