--- src/auth/fake_cega/users/homer.yml
@@ -1,2 +1,0 @@
----
-password_hash: $1$bla$EeLIAvDpa2Bwg/S2jqn3T1
