--- src/auth/fake_cega/users/juha.yml
@@ -1,3 +1,0 @@
----
-password_hash: $1$xyz$OcPwHHMV7Y2fEaYljaqOX/
-expiration: INTERVAL '3' MONTH
