--- src/auth/fake_cega/users/santa.yml
@@ -1,2 +1,0 @@
----
-password_hash: $1$xyz$BuJSZKSSNzxpx1.erEcp21
